rename DetailsModal to details-modal
