+++ packages/frinx-workflow-ui/src/pages/workflow-list/workflow-exec/details-modal/details-modal.tsx
